+++ src/Glutinum.Converter/Prelude.fs
@@ -1,0 +1,4 @@
+module Glutinum.Converter.Prelude
+
+[<Literal>]
+let VERSION = "0.4.0"